MAJOR: change CalcTheta to logspace
Original code creates a linspace in theta, so angular spacing is uniform. However,
this results in non-uniform speed spacing, neither linear nor log scaled. In
particular, speeds are concentrated near the center of the speed range, and this
distribution changes for different dxdt values. This commit makes the spacing
uniform in the log-domain and independent of dxdt. `pcolor` replaces `imagesc` in
PlotRadon.m because the theta spacing is no longer linear.